+++ highlight.io/highlight.config.ts
@@ -5,19 +5,3 @@ export const withPageRouterHighlight = PageRouterHighlight({
 	debug: false,
 	serviceName: 'highlight.io',
 })
-
-import pino from 'pino'
-export const logger = pino({
-	transport: {
-		targets: [
-			{
-				target: '@highlight-run/pino',
-				options: {
-					projectID: '4d7k1xeo',
-					serviceName: 'highlight.io-pino',
-				},
-				level: 'trace',
-			},
-		],
-	},
-})
